Fix: SrotageManager 오버로딩
